backend/requirements.txt
# ─────────────────────────────
# 🌐 Core Framework
# ─────────────────────────────
fastapi==0.114.2
uvicorn==0.30.6
starlette==0.38.6
python-multipart==0.0.12
anyio==4.11.0
h11==0.16.0
sniffio==1.3.1

# ─────────────────────────────
# 🧠 Data & Machine Learning
# ─────────────────────────────
numpy==1.26.4          # ✅ compatible with SciPy 1.14.x and pandas 2.2.x
pandas==2.2.3
scikit-learn==1.5.2
scipy==1.14.1
joblib==1.4.2

# ─────────────────────────────
# 🔐 Authentication & Security
# ─────────────────────────────
bcrypt==4.1.3
passlib==1.7.4
PyJWT==2.9.0
SQLAlchemy==2.0.36

# ─────────────────────────────
# 🧰 Utilities
# ─────────────────────────────
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
click==8.3.0
colorama==0.4.6
typing-inspection==0.4.2
typing_extensions==4.12.2
setuptools==80.9.0
wheel==0.45.1
pydantic==2.9.2
pydantic-settings==2.4.0

# ─────────────────────────────
# 📄 Document & File Processing
# ─────────────────────────────
openpyxl==3.1.5
python-docx==1.2.0
PyPDF2==3.0.1
pdfminer.six==20231228
pdfplumber==0.11.4
mammoth==1.6.0
odfpy==1.4.1
python-json-logger==2.0.7
reportlab==4.2.5

# ─────────────────────────────
# ☁️ Environment & Cloud Integration
# ─────────────────────────────
python-dotenv==1.0.1
supabase==2.5.1
postgrest==0.16.6
gotrue==2.6.1
storage3==0.7.6

# ─────────────────────────────
# 📦 Networking & File Handling
# ─────────────────────────────
filetype==1.2.0
charset-normalizer==3.3.2
requests==2.32.3   # ✅ Added to fix verify_stack.py import

# ─────────────────────────────
# ✅ Logging & Diagnostics
# ─────────────────────────────
python-json-logger==2.0.7
